+++ provider/openstack/package_test.go
@@ -1,0 +1,26 @@
+// Copyright 2012-2015 Canonical Ltd.
+// Licensed under the AGPLv3, see LICENCE file for details.
+
+package openstack_test
+
+import (
+	"flag"
+	"testing"
+
+	gc "gopkg.in/check.v1"
+	"launchpad.net/goose/identity"
+)
+
+var live = flag.Bool("live", false, "Include live OpenStack tests")
+
+func Test(t *testing.T) {
+	if *live {
+		cred, err := identity.CompleteCredentialsFromEnv()
+		if err != nil {
+			t.Fatalf("Error setting up test suite: %s", err.Error())
+		}
+		registerLiveTests(cred)
+	}
+	registerLocalTests()
+	gc.TestingT(t)
+}
